Google map integrated for RestRooms and Api call successful

+++ app/src/main/java/com/fame/plumbum/chataround/restroom/presenter/RestRoomPresenter.java
@@ -6,7 +6,7 @@ package com.fame.plumbum.chataround.restroom.presenter;
 
 public interface RestRoomPresenter {
 
-    void requestRestRooms(Float latitude, Float langitude);
+    void requestRestRooms(Double latitude, Double langitude);
 
 
 }
